refactor: 날짜 포맷 위치 변경
