+++ app/src/test/java/woowacourse/payments/CardsTest.kt
@@ -5,6 +5,7 @@ import org.junit.jupiter.api.Test
 import org.junit.jupiter.api.assertNotNull
 import org.junit.jupiter.api.assertThrows
 import woowacourse.payments.domain.Card
+import java.time.YearMonth
 
 class CardsTest {
     @Test
@@ -13,7 +14,7 @@ class CardsTest {
         val result =
             Card.create(
                 cardNumber = "뭉치바보",
-                expiryDate = "1234",
+                expiryDate = YearMonth.of(2034, 12),
                 cardOwner = "뭉치",
                 password = "1234",
             )
@@ -31,7 +32,7 @@ class CardsTest {
         val result =
             Card.create(
                 cardNumber = cardNumber,
-                expiryDate = "1234",
+                expiryDate = YearMonth.of(2034, 12),
                 cardOwner = "뭉치",
                 password = "1234",
             )
@@ -43,29 +44,12 @@ class CardsTest {
     }
 
     @Test
-    fun `만료일 변환실패 시 예외가 발생한다`() {
-        // given
-        val result =
-            Card.create(
-                cardNumber = "1234567812345678",
-                expiryDate = "뭉치바보",
-                cardOwner = "뭉치",
-                password = "1234",
-            )
-
-        // when & then
-        val exception = assertThrows<IllegalArgumentException> { result.getOrThrow() }
-        assertNotNull(exception.message)
-        Assertions.assertTrue(exception.message!!.contains("만료일 날짜변환 오류"))
-    }
-
-    @Test
     fun `만료일이 현재보다 과거일 시 예외가 발생한다`() {
         // given
         val result =
             Card.create(
                 cardNumber = "1234567812345678",
-                expiryDate = "1224",
+                expiryDate = YearMonth.of(2024, 12),
                 cardOwner = "뭉치",
                 password = "1234",
             )
@@ -82,7 +66,7 @@ class CardsTest {
         val result =
             Card.create(
                 cardNumber = "1234567812345678",
-                expiryDate = "1234",
+                expiryDate = YearMonth.of(2034, 12),
                 cardOwner = "뭉치",
                 password = "뭉치바보",
             )
@@ -99,7 +83,7 @@ class CardsTest {
         val result =
             Card.create(
                 cardNumber = "1234567812345678",
-                expiryDate = "1234",
+                expiryDate = YearMonth.of(2034, 12),
                 cardOwner = "뭉치",
                 password = "12345",
             )
